dia  3

+++ dia 2 soma/sum.py
@@ -1,7 +1,7 @@
 def sum(num1, num2):
     return num1 + num2
 
-num1 = int(input("Digite um número: "))
-num2 = int(input("Digite outro número: "))
+num1 = float(input("Digite um número: "))
+num2 = float(input("Digite outro número: "))
 
 print ("O resultado da soma é %d" %sum(num1, num2))
